posts/requirements.txt
fastapi
uvicorn
sqlmodel
bcrypt
python-jose[cryptography]
authlib
httpx
itsdangerous
python-multipart
psycopg2-binary  # PostgreSQL driver
python-dotenv    # For .env file support
aiosmtplib==3.0.0
boto3            # For Cloudflare R2 (S3-compatible) uploads
pillow
taskiq
taskiq-redis
redis

# For checker.py
requests
jwt